Move MainPageWithScroll to components/common
