--- airbyte-webapp/src/components/MainPageWithScroll/index.tsx
@@ -1,4 +1,0 @@
-import MainPageWithScroll from "./MainPageWithScroll";
-
-export default MainPageWithScroll;
-export { MainPageWithScroll };
