+++ airbyte-webapp/src/components/common/MainPageWithScroll/MainPageWithScroll.module.scss
@@ -1,4 +1,4 @@
-@use "../../scss/variables";
+@use "scss/variables";
 
 .page {
   overflow-y: hidden;
