+++ airbyte-webapp/src/components/common/MainPageWithScroll/index.ts
@@ -1,0 +1,1 @@
+export * from "./MainPageWithScroll";
