+++ airbyte-webapp/src/packages/cloud/views/credits/CreditsPage/CreditsPage.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { FormattedMessage } from "react-intl";
 
 import { HeadTitle } from "components/common/HeadTitle";
-import MainPageWithScroll from "components/MainPageWithScroll";
+import { MainPageWithScroll } from "components/common/MainPageWithScroll";
 import { PageHeader } from "components/ui/PageHeader";
 
 import { PageTrackingCodes, useTrackPage } from "hooks/services/Analytics";
